+++ src/components/NcTextField/NcTextField.vue
@@ -134,7 +134,7 @@ export default {
 
 		<!-- Trailing icon slot, except for search type input as the browser already adds a trailing close icon -->
 		<template v-if="type !== 'search'" #trailing-button-icon>
-			<NcIconSvgWrapper v-if="isArrow" directional :path="mdiArrowRight" />
+			<NcIconSvgWrapper v-if="trailingButtonIcon === 'arrowEnd'" directional :path="mdiArrowRight" />
 			<NcIconSvgWrapper v-else :path="trailingButtonIcon === 'undo' ? mdiUndo : mdiClose" />
 		</template>
 	</NcInputField>
@@ -153,11 +153,8 @@ import { t } from '../../l10n.js'
 const props = withDefaults(defineProps<NcInputFieldProps & {
 	/**
 	 * Specifies which material design icon should be used for the trailing button.
-	 *
-	 * The value `'arrowRight'` is deprecated and will be removed with the next version.
-	 * Instead use the `arrowEnd` option which reflects that the directions depends on the text directions (LTR vs RTL).
 	 */
-	trailingButtonIcon?: 'arrowEnd' | 'arrowRight' | 'close' | 'undo'
+	trailingButtonIcon?: 'arrowEnd' | 'close' | 'undo'
 
 	/**
 	 * The `aria-label` to set on the trailing button
@@ -193,25 +190,11 @@ defineSlots<{
 
 const inputField = useTemplateRef('inputField')
 
-/**
- * Is the trailing button icon directional.
- * Meaning the icon needs to be flipped on RTL text flow.
- */
-const isArrow = computed(() => props.trailingButtonIcon === 'arrowEnd' || props.trailingButtonIcon === 'arrowRight')
-
-const defaultTrailingButtonLabel = computed(() => {
-	switch (props.trailingButtonIcon) {
-	case 'arrowEnd':
-	case 'arrowRight':
-		return t('Save changes')
-	case 'close':
-		return t('Clear text')
-	case 'undo':
-		return t('Undo changes')
-	default:
-		return undefined
-	}
-})
+const defaultTrailingButtonLabels = {
+	arrowEnd: t('Save changes'),
+	close: t('Clear text'),
+	undo: t('Undo changes'),
+}
 
 const NcInputFieldPropNames = new Set(Object.keys(NcInputField.props))
 const propsToForward = computed<NcInputFieldProps>(() => {
@@ -220,7 +203,7 @@ const propsToForward = computed<NcInputFieldProps>(() => {
 			.filter(([key]) => NcInputFieldPropNames.has(key)),
 	)
 
-	sharedProps.trailingButtonLabel ??= defaultTrailingButtonLabel.value
+	sharedProps.trailingButtonLabel ??= defaultTrailingButtonLabels[props.trailingButtonIcon]
 	return sharedProps satisfies NcInputFieldProps
 })